\name{GDPSamp-package}
\alias{GDPSamp-package}
\alias{GDPSamp}
\docType{package}
\title{
\packageTitle{GDPSamp}
}
\description{
\packageDescription{GDPSamp}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{GDPSamp}
\packageIndices{GDPSamp}
An overview of how to use the package, including the most important functions
}
\author{
\packageAuthor{GDPSamp}

Maintainer: \packageMaintainer{GDPSamp}
}
\references{
Literature or other references for background information 
}
 Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory
\keyword{ package }
\seealso{
 Optional links to other man pages, e.g. 
}
